% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_soundSample.R
\name{as.Wave}
\alias{as.Wave}
\title{Cast to a tuneR::Wave object}
\usage{
as.Wave(x)
}
\arguments{
\item{x}{sound sample object.}
}
\value{
a tuneR Wave object.
}
\description{
Convert a soundSample into a tuneR::Wave object.
}
\examples{
sam <- soundSample(sin(2*pi*seq(0,1,,44100)*440)) # 1-second A (440 Hz)
w <- as.Wave(sam)
tuneR::plot(w)
}
