% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bell}
\alias{bell}
\title{Bell sample}
\format{
An object of class \code{soundSample} of length 4.
}
\usage{
bell
}
\description{
A ride cymbal (hit ion the bell) sound sample object
}
\keyword{datasets}
