% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{checkMaxSize}
\alias{checkMaxSize}
\title{Check wave size}
\usage{
checkMaxSize(n, nmax)
}
\arguments{
\item{n}{integer, size to be checked}

\item{nmax}{integer, maximum allowed size}
}
\value{
nothing - just stops execution with an error message if n>nmax
}
\description{
Check that the size of a wave does not exceed the maximum allowed size.
}
