% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{checkSeqArgs}
\alias{checkSeqArgs}
\title{Check sequencer arguments
Check that the arguments used in sequencing functions (e.g. time, volume, pan, etc.) are valid.}
\usage{
checkSeqArgs(argList)
}
\arguments{
\item{argList}{list, a named list containg the arguments}
}
\value{
nothing - just stops execution with an error message if something is invalid
}
\description{
Check sequencer arguments
Check that the arguments used in sequencing functions (e.g. time, volume, pan, etc.) are valid.
}
