% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effects.R
\name{disto_clip}
\alias{disto_clip}
\title{Clip distortion}
\usage{
disto_clip(x, level)
}
\arguments{
\item{x}{Numeric vector in [-1,1], input signal}

\item{level}{Numeric (>=0), distortion level}
}
\value{
a numeric vector containing the distorted output signal
}
\description{
Transfer function for 'clip' distortion
}
