% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effects.R
\name{disto_tanh}
\alias{disto_tanh}
\title{Tanh distortion}
\usage{
disto_tanh(x, level)
}
\arguments{
\item{x}{Numeric vector in [-1,1], input signal}

\item{level}{Numeric (>=0), distortion level}
}
\value{
a numeric vector containing the distorted output signal
}
\description{
Transfer function for 'tanh' distortion
}
