% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessibility.R
\name{.accessible_pos}
\alias{.accessible_pos}
\title{find nonzero pos in the results of the conv}
\usage{
.accessible_pos(dat, window, rm_overlap)
}
\description{
find nonzero pos in the results of the conv
}
