% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessibility.R
\name{.forward}
\alias{.forward}
\title{forward pass to perform convolution}
\usage{
.forward(x, weights, window)
}
\arguments{
\item{x}{the vector of data to find sequences in}

\item{weights}{the weights tensor to be passed to the conv layer}

\item{window}{the window size}
}
\description{
forward pass to perform convolution
}
