% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessibility.R
\name{.generate_weights}
\alias{.generate_weights}
\title{function to generate the weight vector
for the convolution step}
\usage{
.generate_weights(window, continuous, required_pad)
}
\arguments{
\item{window}{the window size}

\item{continuous}{bool, continuous or non-continuous}

\item{required_pad}{the padding for non-continuous windows}
}
\description{
function to generate the weight vector
for the convolution step
}
