% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessibility.R
\name{find_sequences}
\alias{find_sequences}
\title{Find sequences of a given window size
in a given vector of 1s and 0s}
\usage{
find_sequences(
  data,
  window = 1,
  continuous = TRUE,
  required_pad = 1,
  rm_overlap = FALSE
)
}
\arguments{
\item{data}{1d vector of 1s and 0s}

\item{window}{size of sequence}

\item{continuous}{sequence of continuous 1s or
1s either end with 0s between}

\item{required_pad}{if non-continuous how many 1s at either end}

\item{rm_overlap}{if TRUE find earliest sequence, else return all possible sequences}
}
\value{
vector of indices were criteria is met
}
\description{
Find sequences of a given window size
in a given vector of 1s and 0s
}
