% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthesizers.R
\name{getSynthNote}
\alias{getSynthNote}
\title{Single note from a synthesizer}
\usage{
getSynthNote(
  freq,
  nHarmonics = 5,
  peak = 0.03,
  decay = 0.8,
  duration = 1,
  sustain = 0.25,
  decayPar = 1,
  sustainPar = 4,
  type = "sine"
)
}
\arguments{
\item{freq}{Numeric, base frequency in Hz}

\item{nHarmonics}{Integer >=1, number of harmonics}

\item{peak}{Numeric, peak time in seconds}

\item{decay}{Numeric, end-of-decay time in seconds}

\item{duration}{Numeric, total duration in seconds}

\item{sustain}{Numeric, sustain volume}

\item{decayPar}{Numeric, the higher the value the smaller the decay time for higher harmonics}

\item{sustainPar}{Numeric, the higher the value the smaller the sustain volume for higher harmonics}

\item{type}{String, oscillator type, one of 'sine', 'saw', 'square' or 'triangle'.
If an unknown string is provided, a sine oscillator will be used.}
}
\value{
An object of class 'soundSample'.
}
\description{
Creates one note with frequency freq from an additive, Hammond-inspired synth.
Higher harmonics decay faster and have smaller sustain.
}
\examples{
sam <- getSynthNote(440,nHarmonics=7)
plot(sam)
\dontrun{
# This line of code is wrapped in \dontrun{} since it relies
# on an external audio player to listen to the audio sample.
# See ?tuneR::setWavPlayer for setting a default player.
listen(sam)}
}
