% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_soundSample.R
\name{getTime}
\alias{getTime}
\title{Get sampling time}
\usage{
getTime(x)
}
\arguments{
\item{x}{sound sample object.}
}
\value{
a numeric vector containing the sampling times in second.
}
\description{
Get the times steps associated with a sound sample.
}
\examples{
# Define sound sample
sam <- soundSample(sin(2*pi*seq(0,1,,44100)*440)+0.1*rnorm(44100)) # 1-second noisy A
# Compute sampling times
timeSteps=getTime(sam)
}
