% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{globalT}
\alias{globalT}
\title{Global Temperature Anomalies dataset}
\format{
An object of class \code{data.frame} with 172 rows and 2 columns.
}
\source{
\url{https://www.metoffice.gov.uk/hadobs/hadcrut4/data/current/download.html}
}
\usage{
globalT
}
\description{
Times series of annual temperature anomalies at the global scale,
1850-2021. This time series is the one used to create the Warming Stripes
(\url{https://www.climate-lab-book.ac.uk/2018/warming-stripes/}).
}
\keyword{datasets}
