% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hiHat2}
\alias{hiHat2}
\title{Hi-hat sample 2}
\format{
An object of class \code{soundSample} of length 4.
}
\source{
\url{http://www.archive.org/details/OpenPathMusic44V1}
}
\usage{
hiHat2
}
\description{
A hi-hat sound sample object
}
\keyword{datasets}
