% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hiHat_o}
\alias{hiHat_o}
\title{Open Hi-hat sample}
\format{
An object of class \code{soundSample} of length 4.
}
\usage{
hiHat_o
}
\description{
An open hi-hat sound sample object
}
\keyword{datasets}
