% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_instrument.R
\name{instrument}
\alias{instrument}
\title{Instrument constructor.}
\usage{
instrument(samples, notes = as.character(1:length(samples)))
}
\arguments{
\item{samples}{list of sound samples}

\item{notes}{string vector, name given to each sample}
}
\value{
An object of class 'Instrument'.
}
\description{
Creates a new instance of an 'instrument' object.
An instrument is a named list of sound samples (all with the same sampling rate).
}
\examples{
drumset <- instrument(samples=list(kick,snare,hiHat),notes=c('boom','tat','cheet'))
}
