% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{kick2}
\alias{kick2}
\title{Kick sample2}
\format{
An object of class \code{soundSample} of length 4.
}
\source{
\url{http://www.archive.org/details/OpenPathMusic44V1}
}
\usage{
kick2
}
\description{
A kick sound sample object
}
\keyword{datasets}
