% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_soundSample.R
\name{listen}
\alias{listen}
\title{Listen to a sound sample}
\usage{
listen(x)
}
\arguments{
\item{x}{sound sample object.}
}
\value{
nothing - listening function.
}
\description{
Listen to a sound sample. Based on tuneR function 'play'
}
\examples{
# Define sound sample
sam <- soundSample(sin(2*pi*seq(0,1,,44100)*440)+0.1*rnorm(44100)) # 1-second noisy A
# Uncomment to play it
# listen(sam)
}
