% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mini909}
\alias{mini909}
\title{TR-909 minimalistic drumkit}
\format{
An object of class \code{instrument} of length 6.
}
\source{
\url{https://freesound.org/people/altemark/packs/1643/}
}
\usage{
mini909
}
\description{
An instrument containing a few basic sounds from a TR-909-inspired drumkit
}
\keyword{datasets}
