% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oscillators.R
\name{oscillator}
\alias{oscillator}
\title{General oscillator}
\usage{
oscillator(type = "sine", freq = 440, duration = 1, phase = 0, rate = 44100)
}
\arguments{
\item{type}{String, oscillator type, one of 'sine', 'saw', 'square' or 'triangle'.
If an unknown string is provided, a sine oscillator will be used.}

\item{freq}{Numeric, note frequency in Hz}

\item{duration}{Numeric, note duration in second}

\item{phase}{Numeric, phase in radians (typically between 0 and 2*pi)}

\item{rate}{Numeric, sampling rate in Hz}
}
\value{
An object of class 'soundSample'.
}
\description{
Creates a soundSample using a oscillator.
}
\examples{
sam <- oscillator(type='saw',freq=220,duration=0.1)
plot(sam)
\dontrun{listen(sam)}
}
