% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_envelope.R
\name{plot.envelope}
\alias{plot.envelope}
\title{Plot}
\usage{
\method{plot}{envelope}(x, ...)
}
\arguments{
\item{x}{envelope object.}

\item{...}{further arguments passed to the base plot function.}
}
\value{
nothing - plotting function.
}
\description{
Plot an envelope.
}
\examples{
# Define envelope
env <- envelope(t=c(0,0.1,0.3,0.8,1),v=c(0,1,0.4,0.4,0))
# plot it
plot(env)
}
