% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_soundSample.R
\name{plot.soundSample}
\alias{plot.soundSample}
\title{Plot a sound sample}
\usage{
\method{plot}{soundSample}(x, ...)
}
\arguments{
\item{x}{sound sample object.}

\item{...}{further arguments passed to tuneR plotting function.}
}
\value{
nothing - plotting function.
}
\description{
Plot a sound sample. Uses plotly to add zooming capability.
}
\examples{
# Define sound sample
sam <- soundSample(sin(2*pi*seq(0,1,,44100)*440)+0.1*rnorm(44100)) # 1-second noisy A
# plot it
plot(sam)
}
