% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_soundSample.R
\name{read.soundSample}
\alias{read.soundSample}
\title{Read a sound sample}
\usage{
read.soundSample(file, ...)
}
\arguments{
\item{file}{string, file with extension .wav or.mp3}

\item{...}{additional arguments passed to function tuneR::readWave}
}
\value{
An object of class 'soundSample'.
}
\description{
Read a sound sample from a .mp3 or .wav file.
}
\examples{
sam=try(read.soundSample(file='vignettes/07027201.mp3'))
}
