% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_manipulation.R
\name{seq_add_period_column}
\alias{seq_add_period_column}
\title{Adds a 'Current Period', 'Previous Period', 'Last Year' column toa table containing dates}
\usage{
seq_add_period_column(
  data,
  end_date,
  window_length = 7,
  periodicity = "day",
  date_col = "date",
  drop_na = TRUE
)
}
\arguments{
\item{drop_na}{}
}
\description{
Adds a 'Current Period', 'Previous Period', 'Last Year' column toa table containing dates
}
