% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_manipulation.R
\name{seq_difference_table}
\alias{seq_difference_table}
\title{Converts a KPI table into period on period and year on year comparisons}
\usage{
seq_difference_table(data, date_last, date_window_length)
}
\arguments{
\item{date_window_length}{}
}
\description{
Converts a KPI table into period on period and year on year comparisons
}
