% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot.R
\name{seq_ggplot_minimal}
\alias{seq_ggplot_minimal}
\title{Standard Sequence formatting for a plotly chart}
\usage{
seq_ggplot_minimal()
}
\description{
Standard Sequence formatting for a plotly chart
}
\examples{
mtcars |>
ggplot2::ggplot(ggplot2::aes(x = disp, y = hp)) +
  ggplot2::geom_line() +
  seq_ggplot_minimal() +
  ggplot2::labs(title = "Title",
                subtitle = "Subtitle",
                caption = "source: a source")

}
