% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{seq_google_login_screen}
\alias{seq_google_login_screen}
\title{Google login screen with a background image}
\usage{
seq_google_login_screen(req, title = "App Login")
}
\arguments{
\item{bg_image}{}
}
\description{
Google login screen with a background image
}
\examples{
# When calling shiny use...
# shinyApp(gar_shiny_ui(ui, login_ui = seq_google_login_screen()), server)
}
