% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\name{seq_heatscale}
\alias{seq_heatscale}
\title{Get a heat colour ramp}
\usage{
seq_heatscale(n = 10)
}
\arguments{
\item{n}{Number of colours to return}
}
\value{
A vector of hex codes
}
\description{
Get a heat colour ramp
}
\examples{
seq_heatscale(12)
}
