% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\name{seq_pal_numeric_no_na}
\alias{seq_pal_numeric_no_na}
\title{A scales::col_numeric() palette with colours for extreme values instead of NA}
\usage{
seq_pal_numeric_no_na(x)
}
\arguments{
\item{x}{}
}
\description{
A scales::col_numeric() palette with colours for extreme values instead of NA
}
