% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\docType{data}
\name{seq_palette}
\alias{seq_palette}
\title{Access the Sequence colour palette}
\format{
An object of class \code{list} of length 7.
}
\usage{
seq_palette
}
\value{
A named list of colours
}
\description{
Access the Sequence colour palette
}
\examples{
seq_palette
}
\keyword{datasets}
