% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotly.R
\name{seq_plotly_minimal}
\alias{seq_plotly_minimal}
\title{Apply minimal formatting to a plotly chart}
\usage{
seq_plotly_minimal(plot, labels_colour = NA)
}
\arguments{
\item{plot}{The plot to which formatting will be applied}

\item{labels_colour}{A colour to use for axis text. Defaults to seq_palette()$darkgrey}
}
\value{
A plotly plot
}
\description{
Apply minimal formatting to a plotly chart
}
\examples{
plotly::plot_ly() |>
  seq_plotly_minimal()
}
