% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{seq_shiny_bslib_theme}
\alias{seq_shiny_bslib_theme}
\title{Sets a Sequence theme for bslib}
\usage{
seq_shiny_bslib_theme()
}
\value{
bslib theme
}
\description{
Sets a Sequence theme for bslib
}
\examples{
# within app ui...
# theme = seq_shiny_bslib_theme()
}
