% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{snare2}
\alias{snare2}
\title{Snare sample 2}
\format{
An object of class \code{soundSample} of length 4.
}
\source{
\url{http://www.archive.org/details/OpenPathMusic44V1}
}
\usage{
snare2
}
\description{
A snare sound sample object
}
\keyword{datasets}
