% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{WaggaWagga}
\alias{WaggaWagga}
\title{Wagga-Wagga dataset}
\format{
An object of class \code{data.frame} with 79 rows and 3 columns.
}
\source{
\url{http://www.bom.gov.au/cgi-bin/climate/hqsites/site_data.cgi?period=annual&variable=meanT&station=072150}

\url{http://www.bom.gov.au/cgi-bin/climate/hqsites/site_data.cgi?period=annual&variable=rain&station=072150}
}
\usage{
WaggaWagga
}
\description{
Times series of monthly temperatures and precipitations
 recorded at Wagga-Wagga, New South Wales, Australia, 1940-2018
}
\keyword{datasets}
