% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_instrument.R
\name{write.instrument}
\alias{write.instrument}
\title{Write an instrument to file}
\usage{
write.instrument(inst, dir = tempdir(), fmt = "wav")
}
\arguments{
\item{inst}{Instrument object.}

\item{dir}{String, directory where files should be written.}

\item{fmt}{String, 'wav' or 'mp3'.}
}
\value{
nothing - writing function.
}
\description{
Write each sound sample of the instrument as a separate .wav or .mp3 file.
}
\examples{
# Create an instrument
drumset <- instrument(samples=list(kick,snare,hiHat),notes=c('boom','tat','cheet'))
# Write to files (one per element)
write.instrument(drumset)
}
