% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_soundSample.R
\name{write.soundSample}
\alias{write.soundSample}
\title{Write a sound sample}
\usage{
write.soundSample(x, file)
}
\arguments{
\item{x}{sound sample object.}

\item{file}{string, destination file. Default file format is .wav.
If file extension is .mp3, conversion to mp3 is attempted using ffmpeg,
which hence needs to be available (see https://ffmpeg.org/).}
}
\value{
nothing - writing function.
}
\description{
Write a sound sample in .wav or .mp3 format.
}
\examples{
sam <- soundSample(sin(2*pi*seq(0,1,,44100)*440)) # 1-second A (440 Hz)
write.soundSample(sam,'A440.wav')
}
