% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{getTime}
\alias{getTime}
\title{getTime function
Compute the time vector starting from 0 associated with a duration and a sampling rate}
\usage{
getTime(duration = 1, rate = 44100)
}
\arguments{
\item{duration}{Numeric}

\item{rate}{Numeric}
}
\value{
a numeric vector
}
\description{
getTime function
Compute the time vector starting from 0 associated with a duration and a sampling rate
}
