% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{pan2vol}
\alias{pan2vol}
\title{Pan-to-volume function}
\usage{
pan2vol(pan, channel)
}
\arguments{
\item{pan}{Numeric between -1 and 1}
}
\value{
a list of length 2 with fields left and right.
}
\description{
Transforms a pan (between -1 for full left and 1 for full right) into right/left
normalized volumes between 0 and 1
}
\keyword{internal}
