% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\name{seq_palette}
\alias{seq_palette}
\title{Access the Sequence colour palette}
\usage{
seq_palette()
}
\value{
A named list of colours
}
\description{
Access the Sequence colour palette
}
\examples{
seq_palette()
}
